+++ NEWS.md
@@ -1,4 +1,10 @@
-# rAmCharts4 1.4.1
+# rAmCharts4 1.5.0 (2021-10-10)
+
+* Fixed a bug: bullets were unresponsive on zooming.
+* New option "zoom buttons" for the line chart and the scatter chart.
+
+
+# rAmCharts4 1.4.1 (2021-09-30)
 
 * Fixed a bug: `amPercentageBarChart` did not work with Shiny.
 * New function `updateAmPercentageBarChart`, to update the data of a 100%